app/dev-requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile dev-requirements.in
#
aiohttp==3.8.5
    # via
    #   -r requirements.txt
    #   openai
aiosignal==1.3.1
    # via
    #   -r requirements.txt
    #   aiohttp
anyio==3.7.1
    # via
    #   -r requirements.txt
    #   httpcore
    #   starlette
async-timeout==4.0.2
    # via
    #   -r requirements.txt
    #   aiohttp
attrs==23.1.0
    # via
    #   -r requirements.txt
    #   aiohttp
beautifulsoup4==4.12.2
    # via
    #   -r requirements.txt
    #   marvin
black==23.7.0
    # via -r dev-requirements.in
build==0.10.0
    # via pip-tools
certifi==2023.7.22
    # via
    #   -r requirements.txt
    #   httpcore
    #   httpx
    #   requests
cfgv==3.3.1
    # via pre-commit
charset-normalizer==3.2.0
    # via
    #   -r requirements.txt
    #   aiohttp
    #   requests
click==8.1.6
    # via
    #   -r requirements.txt
    #   black
    #   pip-tools
    #   typer
    #   uvicorn
conventional-pre-commit==2.3.0
    # via -r dev-requirements.in
distlib==0.3.7
    # via virtualenv
fastapi==0.100.1
    # via
    #   -r requirements.txt
    #   marvin
filelock==3.12.2
    # via virtualenv
frozenlist==1.4.0
    # via
    #   -r requirements.txt
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via
    #   -r requirements.txt
    #   httpcore
    #   uvicorn
httpcore==0.17.3
    # via
    #   -r requirements.txt
    #   httpx
httpx==0.24.1
    # via
    #   -r dev-requirements.in
    #   -r requirements.txt
    #   marvin
identify==2.5.26
    # via pre-commit
idna==3.4
    # via
    #   -r requirements.txt
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via
    #   -r requirements.txt
    #   marvin
jsonpatch==1.33
    # via
    #   -r requirements.txt
    #   marvin
jsonpointer==2.4
    # via
    #   -r requirements.txt
    #   jsonpatch
markdown-it-py==3.0.0
    # via
    #   -r requirements.txt
    #   rich
markupsafe==2.1.3
    # via
    #   -r requirements.txt
    #   jinja2
marvin==1.2.1
    # via -r requirements.txt
mdurl==0.1.2
    # via
    #   -r requirements.txt
    #   markdown-it-py
multidict==6.0.4
    # via
    #   -r requirements.txt
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via black
nodeenv==1.8.0
    # via pre-commit
openai==0.27.8
    # via
    #   -r requirements.txt
    #   marvin
packaging==23.1
    # via
    #   black
    #   build
    #   pytest
pastel==0.2.1
    # via poethepoet
pathspec==0.11.2
    # via black
pip-tools==7.1.0
    # via -r dev-requirements.in
pipdeptree==2.11.0
    # via -r dev-requirements.in
platformdirs==3.10.0
    # via
    #   black
    #   virtualenv
pluggy==1.2.0
    # via pytest
poethepoet==0.21.1
    # via -r dev-requirements.in
pre-commit==3.3.3
    # via -r dev-requirements.in
pydantic[dotenv]==1.10.12
    # via
    #   -r dev-requirements.in
    #   -r requirements.txt
    #   fastapi
    #   marvin
pygments==2.15.1
    # via
    #   -r requirements.txt
    #   rich
pyproject-hooks==1.0.0
    # via build
pytest==7.3.2
    # via -r dev-requirements.in
python-dotenv==1.0.0
    # via
    #   -r requirements.txt
    #   pydantic
pyyaml==6.0.1
    # via pre-commit
regex==2023.6.3
    # via
    #   -r requirements.txt
    #   tiktoken
requests==2.31.0
    # via
    #   -r requirements.txt
    #   openai
    #   tiktoken
rich==13.5.2
    # via
    #   -r requirements.txt
    #   marvin
sniffio==1.3.0
    # via
    #   -r requirements.txt
    #   anyio
    #   httpcore
    #   httpx
soupsieve==2.4.1
    # via
    #   -r requirements.txt
    #   beautifulsoup4
starlette==0.27.0
    # via
    #   -r requirements.txt
    #   fastapi
tiktoken==0.4.0
    # via
    #   -r requirements.txt
    #   marvin
tomli==2.0.1
    # via poethepoet
tqdm==4.65.0
    # via
    #   -r requirements.txt
    #   openai
typer==0.9.0
    # via
    #   -r requirements.txt
    #   marvin
typing-extensions==4.7.1
    # via
    #   -r requirements.txt
    #   fastapi
    #   pydantic
    #   typer
tzdata==2023.3
    # via
    #   -r requirements.txt
    #   marvin
urllib3==2.0.4
    # via
    #   -r requirements.txt
    #   requests
uvicorn==0.23.2
    # via
    #   -r dev-requirements.in
    #   -r requirements.txt
    #   marvin
virtualenv==20.24.2
    # via pre-commit
wheel==0.41.0
    # via pip-tools
yarl==1.9.2
    # via
    #   -r requirements.txt
    #   aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
